+++ src/access/domain/dtos/GuestbookResponseDTO.ts
@@ -1,0 +1,10 @@
+export interface GuestbookAnswerDTO {
+  id: number | string
+  value: string | string[]
+}
+
+export interface GuestbookResponseDTO {
+  guestbookResponse: {
+    answers: GuestbookAnswerDTO[]
+  }
+}